remove webhook tests and unnecessary abstractions, make just one index.py file
