--- infra/scripts/webhook
@@ -1,13 +1,0 @@
-#!/bin/bash
-
-if [ $1 == "run" ]; then
-elif [ $1 == "configure" ]; then
-  # updates the lambda code with ./infra/webhook/function/webhook.py
-  # ensures github has the correct webhook in place (no update if not required)
-elif [ $1 == "pause" ]; then
-
-elif [ $1 == "resume" ]; then
-else
-  print "unknown command $1"
-  exit 1
-fi
