--- infra/webhook/events/sample_push.json
@@ -1,12 +1,0 @@
-{
-  "sample": true,
-  "headers": {
-    "x-github-event": "push"
-  },
-  "body": {
-    "ref": "refs/heads/dci#84",
-    "head_commit": {
-      "id": "90aa35428f689dbeff1a59b010be25420fa6fdd4"
-    }
-  }
-}
